% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredDrugInducedLiverInjury.R
\docType{package}
\name{PredDrugInducedLiverInjury}
\alias{PredDrugInducedLiverInjury}
\title{PredDrugInducedLiverInjury}
\description{
A package for running prediction model development studies using data in the OMOP CDM
}
